Consolidate core redis with feature flag
